backend/requirements.txt
fastapi>=0.115
uvicorn[standard]>=0.30
pydantic>=2.8
SQLAlchemy>=2.0
psycopg[binary]>=3.2
alembic>=1.13
python-dotenv>=1.0
numpy==2.1.2
pillow==10.4.0